ISDB-298: Implementation of c100 validation JUnit test
