+++ src/test/java/uk/gov/hmcts/reform/bulkscan/services/BulkScanC100ServiceTest.java
@@ -119,11 +119,6 @@ class BulkScanC100ServiceTest {
 
     private static final String C100_VALIDATION_REQUEST_PATH =
             "classpath:request/bulk-scan-c100-validate-input.json";
-    private static final String C100_SECTION6B_ERROR_VALIDATION_REQUEST_PATH =
-            "classpath:request/bulk-scan-c100-section6b-error-validate-input.json";
-
-    private static final String C100_SECTION_6B_VALIDATION_REQUEST_PATH =
-            "classpath:request/bulk-scan-c100-section6b-validate-input.json";
 
     private static final String C100_TRANSFORM_REQUEST_PATH =
             "classpath:request/bulk-scan-c100-transform-input.json";
@@ -148,18 +143,6 @@ class BulkScanC100ServiceTest {
     }
 
     @Test
-    @DisplayName("C100 section 6b validation success.")
-    void testC100Section6bValidationSuccess() throws IOException {
-        BulkScanValidationRequest bulkScanValidationRequest =
-                mapper.readValue(
-                        readFileFrom(C100_SECTION_6B_VALIDATION_REQUEST_PATH),
-                        BulkScanValidationRequest.class);
-        BulkScanValidationResponse res =
-                bulkScanValidationService.validate(bulkScanValidationRequest);
-        assertEquals(Status.SUCCESS, res.status);
-    }
-
-    @Test
     @DisplayName("Test a sample successful C100 application validation with full details.")
     void testC100Success() {
         BulkScanValidationRequest bulkScanValidationRequest =
@@ -170,20 +153,6 @@ class BulkScanC100ServiceTest {
         BulkScanValidationResponse res =
                 bulkScanValidationService.validate(bulkScanValidationRequest);
         assertEquals(Status.SUCCESS, res.status);
-    }
-
-    @Test
-    @DisplayName(
-            "Test a sample successful C100 section 6b application validation error with full"
-                    + " details.")
-    void testC100Section6bError() throws IOException {
-        BulkScanValidationRequest bulkScanValidationRequest =
-                mapper.readValue(
-                        readFileFrom(C100_SECTION6B_ERROR_VALIDATION_REQUEST_PATH),
-                        BulkScanValidationRequest.class);
-        BulkScanValidationResponse res =
-                bulkScanValidationService.validate(bulkScanValidationRequest);
-        assertEquals(Status.ERRORS, res.status);
     }
 
     @Test
